GNG-466: Move Editor component to /pages/Editor
